refactor(api): move ComplementaryCertificationCourse.js to certif/shared as it is linked to shared CertificationCourse model
